Add subclass Storage#ReadTaskResult
Storage::readTasks needs to return both list of Tasks and list of
malformed row indices.

This allows us to have list of valid Tasks while reporting malformed
rows to user.

Let's introduce ReadTaskResult that represents the result of the
function Storage::readTasks.

The subclass contains both a list of Tasks and list of malformed
row indices for class Uxie to process.